requirements.txt
# Core dependencies for PubMed Clinical Trial Classifier

# LLM and Structured Output
ollama>=0.1.0                 # Ollama client for local model access
outlines>=0.0.40              # Constrained generation for guaranteed JSON structure
json-repair>=0.7.0            # Fallback JSON repair for malformed outputs

# Data Models and Validation
pydantic>=2.0.0               # Data validation and type safety
pydantic-settings>=2.0.0      # Settings management

# Data Processing
pandas>=2.0.0                 # CSV handling and data manipulation
numpy>=1.24.0                 # Numerical operations

# Configuration and Templates
pyyaml>=6.0.0                 # YAML parsing for prompt templates

# CLI and User Interface
click>=8.0.0                  # CLI framework
rich>=13.0.0                  # Rich terminal output and formatting
tqdm>=4.65.0                  # Progress bars

# Utilities
python-dotenv>=1.0.0          # Environment variable management
loguru>=0.7.0                 # Better logging

# PubMed Scraper dependencies
biopython>=1.81               # Provides Entrez utilities for PubMed API

# Web Scrapers
httpx>=0.25                   # HTTP client with retry support
aiohttp>=3.9                  # Async HTTP client
tenacity>=8.2                 # Retry logic with exponential backoff

# Neo4j Graph Database
neo4j>=5.14                   # Official Neo4j driver

# Author Network Extractor
geopy>=2.4                    # Geocoding
thefuzz>=0.20                 # Fuzzy string matching
networkx>=3.2                 # Network analysis (backup/testing)

# Testing
pytest>=7.4                   # Testing framework
pytest-asyncio>=0.21          # Async testing support
pytest-mock>=3.12             # Mocking utilities
pytest-cov>=4.1.0             # Test coverage reporting
respx>=0.20                   # HTTP mocking for httpx
